# docs/ui-openapi.json
{
  "components": {
    "schemas": {
...
      "EvidencePageView": {
        "additionalProperties": false,
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/RecoveryPlanView"
            },
            "title": "Decisions",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceView"
            },
            "title": "Evidence",
            "type": "array"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "receipts": {
            "items": {
              "$ref": "#/components/schemas/ActionReceiptView"
            },
            "title": "Receipts",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/ExecutionEventView"
            },
            "title": "Timeline",
            "type": "array"
          },
          "verification": {
            "items": {
              "$ref": "#/components/schemas/VerificationView"
            },
            "title": "Verification",
            "type": "array"
          }
        },
        "required": [
          "incident_id",
          "revision",
          "timeline",
          "receipts",
          "verification",
          "decisions",
          "evidence"
        ],
        "title": "EvidencePageView",
        "type": "object"
      },
...
      "ExecutionEventsView": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/ExecutionEventView"
            },
            "title": "Events",
            "type": "array"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "next_cursor": {
            "title": "Next Cursor",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "terminal": {
            "title": "Terminal",
            "type": "boolean"
          }
        },
        "required": [
          "incident_id",
          "revision",
          "events",
          "next_cursor",
          "terminal"
        ],
        "title": "ExecutionEventsView",
        "type": "object"
      },
...
      "ExternalRealityView": {
        "additionalProperties": false,
        "properties": {
          "availability": {
            "default": "EVIDENCE_UNAVAILABLE",
            "enum": [
              "AVAILABLE",
              "EVIDENCE_UNAVAILABLE"
            ],
            "title": "Availability",
            "type": "string"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "resources": {
            "items": {
              "$ref": "#/components/schemas/ExternalResourceView"
            },
            "title": "Resources",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "incident_id",
          "revision"
        ],
        "title": "ExternalRealityView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObjectiveFilter": {
        "enum": [
          "all",
          "active",
          "restored"
        ],
        "title": "ObjectiveFilter",
        "type": "string"
      },
...
      "ObjectivesView": {
        "additionalProperties": false,
        "properties": {
          "filter": {
            "$ref": "#/components/schemas/ObjectiveFilter"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ObjectiveSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "filter",
          "items"
        ],
        "title": "ObjectivesView",
        "type": "object"
      },
...
      "OperatorContextView": {
        "additionalProperties": false,
        "properties": {
          "current_recovery": {
            "$ref": "#/components/schemas/RecoveryAttemptView"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ExecutionEventView"
            },
            "title": "Events",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceView"
            },
            "title": "Evidence",
            "type": "array"
          },
          "objective": {
            "$ref": "#/components/schemas/ObjectiveContext"
          },
          "plans": {
            "items": {
              "$ref": "#/components/schemas/RecoveryPlanView"
            },
            "title": "Plans",
            "type": "array"
          },
          "read_only": {
            "default": true,
            "title": "Read Only",
            "type": "boolean"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "revision",
          "objective",
          "current_recovery",
          "plans",
          "evidence",
          "events"
        ],
        "title": "OperatorContextView",
        "type": "object"
      },
      "OverviewView": {
        "additionalProperties": false,
        "properties": {
          "active_objectives": {
            "items": {
              "$ref": "#/components/schemas/ObjectiveSummary"
            },
            "title": "Active Objectives",
            "type": "array"
          },
          "current_priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurrentPriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "objective_summary": {
            "$ref": "#/components/schemas/ObjectiveCounts"
          },
          "recent_activity": {
            "items": {
              "$ref": "#/components/schemas/ExecutionEventView"
            },
            "title": "Recent Activity",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "current_priority",
          "objective_summary",
          "active_objectives",
          "recent_activity"
        ],
        "title": "OverviewView",
        "type": "object"
      },
...
      "RecoveryCaseView": {
        "additionalProperties": false,
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionReceiptView"
            },
            "title": "Actions",
            "type": "array"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/RecoveryAttemptView"
            },
            "title": "Attempts",
            "type": "array"
          },
          "detect_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DetectContextView"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceView"
            },
            "title": "Evidence",
            "type": "array"
          },
          "objective": {
            "$ref": "#/components/schemas/ObjectiveContext"
          },
          "plans": {
            "items": {
              "$ref": "#/components/schemas/RecoveryPlanView"
            },
            "title": "Plans",
            "type": "array"
          },
          "replan_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReplanContextView"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "summary": {
            "$ref": "#/components/schemas/RecoverySummary"
          },
          "verifications": {
            "items": {
              "$ref": "#/components/schemas/VerificationView"
            },
            "title": "Verifications",
            "type": "array"
          },
          "what_changed": {
            "items": {
              "$ref": "#/components/schemas/AttemptComparisonItem"
            },
            "title": "What Changed",
            "type": "array"
          },
          "world": {
            "$ref": "#/components/schemas/OperationalGraphView"
          }
        },
        "required": [
          "revision",
          "objective",
          "attempts",
          "summary",
          "world",
          "plans",
          "actions",
          "verifications",
          "what_changed",
          "evidence"
        ],
        "title": "RecoveryCaseView",
        "type": "object"
      },
...
      "VerificationView": {
        "additionalProperties": false,
        "properties": {
          "invariants": {
            "items": {
              "$ref": "#/components/schemas/VerificationInvariantView"
            },
            "title": "Invariants",
            "type": "array"
          },
          "objective_id": {
            "title": "Objective Id",
            "type": "string"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "recovery_attempt": {
            "title": "Recovery Attempt",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "verification_id": {
            "title": "Verification Id",
            "type": "string"
          }
        },
        "required": [
          "verification_id",
          "recovery_attempt",
          "objective_id",
          "status",
          "invariants"
        ],
        "title": "VerificationView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Recovery execution plus stable read-only presentation resources.",
    "title": "Reflow Objective Recovery API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ui/evidence/{incident_id}": {
      "get": {
        "operationId": "ui_evidence_api_v1_ui_evidence__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidencePageView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Evidence"
      }
    },
    "/api/v1/ui/objectives": {
      "get": {
        "operationId": "ui_objectives_api_v1_ui_objectives_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ObjectiveFilter",
              "default": "all"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectivesView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Objectives"
      }
    },
    "/api/v1/ui/operator/context": {
      "get": {
        "operationId": "ui_operator_context_api_v1_ui_operator_context_get",
        "parameters": [
          {
            "in": "query",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorContextView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Operator Context"
      }
    },
    "/api/v1/ui/overview": {
      "get": {
        "operationId": "ui_overview_api_v1_ui_overview_get",
        "parameters": [
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Overview"
      }
    },
    "/api/v1/ui/recoveries/{incident_id}": {
      "get": {
        "operationId": "ui_recovery_api_v1_ui_recoveries__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCaseView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Recovery"
      }
    },
    "/api/v1/ui/recoveries/{incident_id}/events": {
      "get": {
        "operationId": "ui_events_api_v1_ui_recoveries__incident_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionEventsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Events"
      }
    },
    "/api/v1/ui/recoveries/{incident_id}/external-reality": {
      "get": {
        "operationId": "ui_external_reality_api_v1_ui_recoveries__incident_id__external_reality_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalRealityView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui External Reality"
      }
    },
...
  }
}